+++ src/index.ts
@@ -6,7 +6,8 @@ import fromZigbee from './converters/fromZigbee';
 import assert from 'assert';
 import * as ota from './lib/ota';
 import allDefinitions from './devices';
-import { Definition, Fingerprint, Zh, OnEventData, OnEventType, Configure, Expose, Tz, OtaUpdateAvailableResult } from './lib/types';
+import * as utils from './lib/utils';
+import { Definition, Fingerprint, Zh, OnEventData, OnEventType, Configure, Expose, Tz, OtaUpdateAvailableResult, KeyValue } from './lib/types';
 import {generateDefinition} from './lib/generateDefinition';
 
 export {
@@ -190,8 +191,24 @@ function prepareDefinition(definition: Definition): Definition {
 
     validateDefinition(definition);
 
+    // Add all the options
     if (!definition.options) definition.options = [];
     const optionKeys = definition.options.map((o) => o.name);
+
+    // Add calibration/precision options based on expose
+    for (const expose of Array.isArray(definition.exposes) ? definition.exposes : definition.exposes(null, null)) {
+        if (!optionKeys.includes(expose.name) && utils.isNumericExposeFeature(expose) && expose.name in utils.calibrateAndPrecisionRoundOptionsDefaultPrecision) {
+            // Battery voltage is not calibratable
+            if (expose.name === 'voltage' && expose.unit === 'mV') continue;
+            const type = utils.calibrateAndPrecisionRoundOptionsIsPercentual(expose.name) ? 'percentual' : 'absolute';
+            definition.options.push(exposes.options.calibration(expose.name, type));
+            if (utils.calibrateAndPrecisionRoundOptionsDefaultPrecision[expose.name] !== 0) {
+                definition.options.push(exposes.options.precision(expose.name));
+            }
+            optionKeys.push(expose.name);
+        }
+    }
+
     for (const converter of [...definition.toZigbee, ...definition.fromZigbee]) {
         if (converter.options) {
             const options = typeof converter.options === 'function' ? converter.options(definition) : converter.options;
@@ -205,6 +222,17 @@ function prepareDefinition(definition: Definition): Definition {
     }
 
     return definition
+}
+
+export function postProcessConvertedFromZigbeeMessage(definition: Definition, payload: KeyValue, options: KeyValue) {
+    // Apply calibration/precision options
+    for (const [key, value] of Object.entries(payload)) {
+        const definitionExposes = Array.isArray(definition.exposes) ? definition.exposes : definition.exposes(null, null);
+        const expose = definitionExposes.find((e) => e.property === key);
+        if (expose?.name in utils.calibrateAndPrecisionRoundOptionsDefaultPrecision && utils.isNumber(value)) {
+            payload[key] = utils.calibrateAndPrecisionRoundOptions(value, options, expose.name);
+        }
+    }
 }
 
 export function addDefinition(definition: Definition) {
@@ -229,8 +257,8 @@ for (const definition of allDefinitions) {
     addDefinition(definition);
 }
 
-export function findByDevice(device: Zh.Device, generateForUnknown: boolean = false) {
-    let definition = findDefinition(device, generateForUnknown);
+export async function findByDevice(device: Zh.Device, generateForUnknown: boolean = false) {
+    let definition = await findDefinition(device, generateForUnknown);
     if (definition && definition.whiteLabel) {
         const match = definition.whiteLabel.find((w) => 'fingerprint' in w && w.fingerprint.find((f) => isFingerprintMatch(f, device)));
         if (match) {
@@ -245,7 +273,7 @@ export function findByDevice(device: Zh.Device, generateForUnknown: boolean = fa
     return definition;
 }
 
-export function findDefinition(device: Zh.Device, generateForUnknown: boolean = false): Definition {
+export async function findDefinition(device: Zh.Device, generateForUnknown: boolean = false): Promise<Definition> {
     if (!device) {
         return null;
     }
@@ -258,7 +286,7 @@ export function findDefinition(device: Zh.Device, generateForUnknown: boolean = 
 
         // Do not add this definition to cache,
         // as device configuration might change.
-        return prepareDefinition(generateDefinition(device));
+        return prepareDefinition((await generateDefinition(device)).definition);
     } else if (candidates.length === 1 && candidates[0].hasOwnProperty('zigbeeModel')) {
         return candidates[0];
     } else {
@@ -289,6 +317,10 @@ export function findDefinition(device: Zh.Device, generateForUnknown: boolean = 
     }
 
     return null;
+}
+
+export async function generateExternalDefinitionSource(device: Zh.Device): Promise<string> {
+    return (await generateDefinition(device)).externalDefinitionSource;
 }
 
 function isFingerprintMatch(fingerprint: Fingerprint, device: Zh.Device) {